remove note functionality

+++ src/main/java/edu/brown/cs/ebwhite/database/TurtleQuery.java
@@ -323,11 +323,12 @@ public class TurtleQuery {
     }
   }
 
-  public static void removeNote(int nodeId) throws SQLException {
+  public static void removeNote(int nodeId, int userId) throws SQLException {
     try (Connection conn = Db.getConnection()) {
       try (PreparedStatement prep = conn
-          .prepareStatement("DELETE FROM notes WHERE id=?;")) {
+          .prepareStatement("DELETE FROM notes WHERE id=? AND userid=?;")) {
         prep.setInt(1, nodeId);
+        prep.setInt(2, userId);
         prep.executeUpdate();
       }
     }
